sarver date/requirements_bot.txt
firebase-admin
google-generativeai
